# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: NSE Stock Market Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gift-nifty/history:
    get:
      operationId: getGiftNiftyHistory
      tags: [giftnifty]
      summary: Get Gift Nifty price history
      description: Returns Gift Nifty historical data for last 3 months
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [1M, 3M, 6M]
            default: 3M
      responses:
        "200":
          description: Gift Nifty history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
...
  /commodities/history:
    get:
      operationId: getCommodityHistory
      tags: [commodities]
      summary: Get commodity price history
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [1M, 3M, 6M]
            default: 3M
      responses:
        "200":
          description: Commodity price history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
  /global-index/quote:
    get:
      operationId: getGlobalIndexQuote
      tags: [global]
      summary: Get global index current quote (Yahoo Finance ~15m delay)
      parameters:
        - name: ticker
          in: query
          required: true
          schema:
            type: string
            enum: ["^NYA", "000001.SS", "^HSI", "^NSEI"]
      responses:
        "200":
          description: Global index quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalIndexQuote"
  /global-index/history:
    get:
      operationId: getGlobalIndexHistory
      tags: [global]
      summary: Get global index price history (Yahoo Finance)
      parameters:
        - name: ticker
          in: query
          required: true
          schema:
            type: string
            enum: ["^NYA", "000001.SS", "^HSI", "^NSEI"]
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [1M, 3M]
            default: 3M
      responses:
        "200":
          description: Global index history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
...
  /nse/quote/{symbol}:
    get:
      operationId: getNseQuote
      tags: [nse]
      summary: Get NSE stock quote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: NSE stock quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NseStock"
  /nse/history/{symbol}:
    get:
      operationId: getNseHistory
      tags: [nse]
      summary: Get NSE stock history
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [1D, 1W, 1M, 3M, 6M, 1Y]
            default: 1M
      responses:
        "200":
          description: NSE stock history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
...
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get portfolio stocks
      responses:
        "200":
          description: Portfolio stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioStock"
    post:
      operationId: addPortfolioStock
      tags: [portfolio]
      summary: Add stock to portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPortfolioStockRequest"
      responses:
        "200":
          description: Added stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioStock"
  /portfolio/{symbol}:
    delete:
      operationId: removePortfolioStock
      tags: [portfolio]
      summary: Remove stock from portfolio
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
  /portfolio/{symbol}/indicators:
    get:
      operationId: getStockIndicators
      tags: [portfolio]
      summary: Get VWAP and RSI indicators for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock indicators
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockIndicators"
  /settings/upstox:
    get:
      operationId: getUpstoxSettings
      tags: [settings]
      summary: Get Upstox API connection status
      responses:
        "200":
          description: Upstox settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstoxSettings"
    post:
      operationId: saveUpstoxSettings
      tags: [settings]
      summary: Save Upstox API credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveUpstoxSettingsRequest"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstoxSettings"
...
  /suggestions/options/{id}/analysis:
    get:
      operationId: getOptionsAnalysis
      tags: [analysis]
      summary: AI Decision Breakdown for one options pick
      description: >
        Multi-timeframe AI explainability for a single options pick, addressed by
        SYMBOL-TYPE-STRIKE (e.g. WIPRO-PE-180). Computed on demand.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Pick identifier in the form SYMBOL-TYPE-STRIKE
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"
        "404":
          description: No analysis available for that pick
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /suggestions/intraday/{symbol}/analysis:
    get:
      operationId: getIntradayAnalysis
      tags: [analysis]
      summary: AI Decision Breakdown for one intraday pick
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"
        "404":
          description: No analysis available for that symbol
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /commodities/{symbol}/analysis:
    get:
      operationId: getCommodityAnalysis
      tags: [analysis]
      summary: AI Decision Breakdown for one commodity
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          description: Yahoo Finance futures symbol, e.g. GC=F
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"
        "404":
          description: No analysis available for that symbol
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
...
  /fo-analyzer:
    post:
      operationId: foAnalyzer
      tags: [insights]
      summary: Analyze F&O trade with AI
      description: Accepts a trade position (CE/PE/FUT) and returns comprehensive AI analysis including direction, confidence, decision, P&L, indicators, insights and intraday chart data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoTradeInput"
      responses:
        "200":
          description: F&O AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
components:
  schemas:
...
    PriceHistory:
      type: object
      properties:
        symbol:
          type: string
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/OHLCV"
      required:
        - symbol
        - period
        - data
    OHLCV:
      type: object
      properties:
        timestamp:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
    NseStock:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        previousClose:
          type: number
        volume:
          type: number
        marketCap:
          type: number
        sector:
          type: string
        pe:
          type: number
        week52High:
          type: number
        week52Low:
          type: number
      required:
        - symbol
        - name
        - price
        - change
        - changePercent
        - open
        - high
        - low
        - previousClose
        - volume
        - marketCap
        - sector
        - pe
        - week52High
        - week52Low
...
    PortfolioStock:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        exchange:
          type: string
        addedAt:
          type: string
        buyPrice:
          type: number
        quantity:
          type: integer
      required:
        - id
        - symbol
        - exchange
        - addedAt
    AddPortfolioStockRequest:
      type: object
      properties:
        symbol:
          type: string
        exchange:
          type: string
        buyPrice:
          type: number
        quantity:
          type: integer
      required:
        - symbol
    StockIndicators:
      type: object
      properties:
        symbol:
          type: string
        price:
          type: number
        vwap:
          type: number
        vwapSignal:
          type: string
          enum: [ABOVE, BELOW, AT]
        vwapDeviation:
          type: number
        rsi:
          type: number
        rsiSignal:
          type: string
          enum: [OVERSOLD, NEUTRAL, OVERBOUGHT]
        smartMoneyFlow:
          type: string
          enum: [ACCUMULATION, DISTRIBUTION, NEUTRAL]
        momentum:
          type: string
          enum: [STRONG_UP, UP, NEUTRAL, DOWN, STRONG_DOWN]
        history:
          type: array
          items:
            $ref: "#/components/schemas/IndicatorPoint"
        updatedAt:
          type: string
      required:
        - symbol
        - price
        - vwap
        - vwapSignal
        - vwapDeviation
        - rsi
        - rsiSignal
        - smartMoneyFlow
        - momentum
        - history
        - updatedAt
    IndicatorPoint:
      type: object
      properties:
        timestamp:
          type: string
        close:
          type: number
        vwap:
          type: number
        rsi:
          type: number
        volume:
          type: number
      required:
        - timestamp
        - close
        - vwap
        - rsi
        - volume
    UpstoxSettings:
      type: object
      properties:
        connected:
          type: boolean
        apiKeyMasked:
          type: string
        clientId:
          type: string
        connectedAt:
          type: string
        liveDataEnabled:
          type: boolean
        hasAccessToken:
          type: boolean
      required:
        - connected
        - liveDataEnabled
...
    SaveUpstoxSettingsRequest:
      type: object
      properties:
        apiKey:
          type: string
        apiSecret:
          type: string
        clientId:
          type: string
        accessToken:
          type: string
      required:
        - apiKey
...
    GlobalIndexQuote:
      type: object
      properties:
        ticker:        { type: string }
        name:          { type: string }
        price:         { type: number }
        change:        { type: number }
        changePercent: { type: number }
        currency:      { type: string }
        dataSource:    { type: string }
      required: [ticker, name, price, change, changePercent, currency]
...
    FoTradeInput:
      type: object
      required: [symbol, optionType, strikePrice, buyPrice, currentPremium, lots, expiry]
      properties:
        symbol:         { type: string, description: "NSE stock or index symbol (RELIANCE, NIFTY, BANKNIFTY)" }
        optionType:     { type: string, enum: [CE, PE, FUT] }
        strikePrice:    { type: number }
        buyPrice:       { type: number, description: "Premium paid per unit when entering" }
        currentPremium: { type: number, description: "Current market premium per unit" }
        lots:           { type: number }
        expiry:         { type: string, description: "Expiry date in YYYY-MM-DD format" }
        entryTime:      { type: string, description: "Entry time in HH:MM format (IST)" }
        stopLoss:       { type: number, description: "Optional stop loss premium level" }
        target:         { type: number, description: "Optional target premium level" }
...
    FoAnalysisResult:
      type: object
      required:
        - symbol
        - optionType
        - strikePrice
        - livePrice
        - liveChange
        - liveChangePct
        - direction
        - confidence
        - decision
        - riskLevel
        - exitSuggestion
        - pnl
        - indicators
        - insights
        - rationale
        - alerts
        - priceChart
        - daysToExpiry
      properties:
        symbol:        { type: string }
        optionType:    { type: string }
        strikePrice:   { type: number }
        livePrice:     { type: number }
        liveChange:    { type: number }
        liveChangePct: { type: number }
        direction:     { type: string, enum: [Bullish, Bearish, Sideways] }
        confidence:    { type: number }
        decision:      { type: string, enum: [HOLD, SELL, BOOK_PROFIT, WAIT, ADD_MORE, AVOID_TRADE] }
        riskLevel:     { type: string, enum: [Low, Medium, High] }
        exitSuggestion:
          $ref: "#/components/schemas/FoExitSuggestion"
        pnl:
          $ref: "#/components/schemas/FoPnl"
        indicators:
          $ref: "#/components/schemas/FoIndicators"
        insights:    { type: array, items: { type: string } }
        rationale:   { type: string }
        alerts:      { type: array, items: { type: string } }
        priceChart:
          type: array
          items:
            $ref: "#/components/schemas/FoChartPoint"
        daysToExpiry: { type: number }

    # ── AI Decision Breakdown ────────────────────────────────────────────
    # One payload shape shared by options, intraday and commodities so the UI
    # components are written once. Market-specific blocks (greeks, optionChain)
    # are nullable rather than duplicated into parallel schemas.
...
    SignalScore:
      type: object
      required: [score, band, signal, categories]
      properties:
        score:
          type: number
          description: 0-100 signal strength. Not a probability of profit.
        band: { type: string }
        signal:
          type: string
          enum: [STRONG_BUY, BUY, WATCH, SELL, STRONG_SELL]
        categories:
          type: array
          items:
            $ref: "#/components/schemas/ScoreCategory"
...
    AnalysisHeadline:
      type: object
      required: [primaryLabel, primaryValue]
      properties:
        primaryLabel:           { type: string }
        primaryValue:           { type: string }
        primaryChangePercent:   { type: number, nullable: true }
        secondaryLabel:         { type: string, nullable: true }
        secondaryValue:         { type: string, nullable: true }
        secondaryChangePercent: { type: number, nullable: true }
    AiAnalysis:
      type: object
      required:
        - id
        - market
        - title
        - subtitle
        - signal
        - signalLabel
        - headline
        - score
        - reasons
        - timeframes
        - multiTimeframe
        - tradePlan
        - risks
        - metrics
        - provenance
      properties:
        id: { type: string }
        market:
          type: string
          enum: [options, intraday, commodities]
        title:    { type: string }
        subtitle: { type: string }
        signal:
          type: string
          enum: [STRONG_BUY, BUY, WATCH, SELL, STRONG_SELL]
        signalLabel: { type: string }
        headline:
          $ref: "#/components/schemas/AnalysisHeadline"
        score:
          $ref: "#/components/schemas/SignalScore"
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisReason"
        timeframes:
          type: array
          items:
            $ref: "#/components/schemas/TimeframeTrend"
        multiTimeframe:
          $ref: "#/components/schemas/MultiTimeframeSummary"
        tradePlan:
          $ref: "#/components/schemas/AnalysisTradePlan"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisRiskFactor"
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisMetric"
        greeks:
          allOf:
            - $ref: "#/components/schemas/AnalysisGreeks"
          nullable: true
        optionChain:
          allOf:
            - $ref: "#/components/schemas/OptionChainContext"
          nullable: true
        provenance:
          $ref: "#/components/schemas/AnalysisProvenance"